Add method to set a Termux RUN_COMMAND on a remote view pending intent
This change allows a remote view in a widget to launch a command in
Termux when clicked. This is useful to start a Termux program on-demand
to update a widget, but let Termux and the program remain closed when
the user is not interacting with the widget.

Add a new protocol method "setRemoteViewRunCommand", that accepts a
command path parameter (which should be a full path to the executable),
a command arguments parameter (which is a comma separated string), and
a run in background parameter (which is a boolean). Store these in an
optional RunCommand data class.

If this optional data class exists for a remote view ID when updating
the widget layout, create a pending intent to invoke the Termux RUN_COMMAND
service and set it on the widget root view.

Below is an example Python script that uses this protocol method to
update a count every time the user taps on the widget.

```
import sys
import termuxgui as tg

with tg.Connection() as c:
    wid = sys.argv[1]
    count = sys.argv[2]
    count += 1

    rv = tg.RemoteViews(c)
    c.send_msg({"method": "setRemoteViewRunCommand", "params:" {
        "rid": rv.rid,
        "path": "/data/data/com.termux/files/usr/bin/python",
        "args": f"widget.py, {wid}, {count}",
        "background", True
    }})

    tv = rv.addTextView()
    rv.setText(tv, f"count: {count}")
    rv.setBackgroundColor(tv, 0xffffffff)
    rv.updateWidget(wid)
```